Update patch set 5
Patch Set 5:

The test this patch added is failing on Travis:
https://travis-ci.org/google/breakpad/builds/156019231

It looks like it's not able to find the new test dmp:
2016-08-29 19:03:41: minidump.cc:4354: ERROR: Minidump could not open minidump ../../src/processor/testdata/linux_stack_pointer_in_stack_alt_name.dmp, error 2: No such file or directory

The test passes for me locally, and the file seems to be in the repository, so I'm not sure what's going on. It's consistently failing all subsequent builds, though, which is unfortunate.

Patch-set: 5