Update patch set 1
Patch Set 1:

Hmm adding +rsesek here.
By looking at the unittest this seems to change the behavior of minidump_stackwalk, at least on windows. Not sure if this can have some side effect on our crash reporting pipeline.

There is one thing I really don't get from the CL description: what problem this is this solving? This description is saying that this is changing the behavior of the processor to match some change in binutils, but doesn't say why.

What I would like to see is either a BUG= or a longer description that explains
what is the *real* problem of the current implementation. 
Do you have an example of a crash that you cannot decode without this CL with the all repro steps?

Patch-set: 1
Reviewer: Primiano Tucci <1002871@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0