Update patch set 1
Patch Set 1:

(4 comments)

My main concern with this is that if you have a memory smashing crash, we could miss out on reports. If something on one thread over-frees or overflows something like the malloc heap, and another thread is also operating on the malloc heap, the "crashing thread" may not be indicative of the actual process state.

Patch-set: 1
Reviewer: Robert Sesek <1144624@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>