Update patch set 2
Patch Set 2:

(4 comments)

You're right that this does introduce cases where we might miss an interesting crash. Conversely, however, all crashes are "interesting" if we don't do this. There's always going to be a thread sitting in a chrome event loop that will pass the principal mapping test. This means that we are currently getting every native crash (heavily sanitized) for every app that loads WebView.

As it stands we would miss these crashes through manual reports, because the feedback processing pipeline applies the same logic of requiring a webview stack frame to be present in the debuggerd output for the crashing thread. We can't decide on the device whether any of the chrome threads are doing anything "interesting", either, I think (although that would be a great way of filtering crashes; maybe that could be done server-side).

Patch-set: 2
Reviewer: Tobias Sargeant <1133706@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>