Update patch set 1
Patch Set 1:

I don’t see HAVE_ARC4RANDOM anywhere outside of ffmpeg in Chromium. If we wanted to push it in from the build, we’d have to modify the .gn files for Breakpad in Chromium when we roll this. Is that how you’d like to proceed?

Patch-set: 1
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>