Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Like we said, we’re just waiting for the answer to one question.

while i'm not easily set up to test whether it's the case that getlimit/setrlimit work within the renderer, the crashes that are being investigated for this bug are happening in the browser process. if these calls are blocked by the sandbox within renderers, will that cause further problems, or would it be harmless?

Patch-set: 1