Update patch set 4
Patch Set 4:

Thanks very much for the feedback. My takeaway is that you will continue to support Windows generally which is a relief. stackwalk isn't so much of an issue for me as we do that server-side (although we are currently compiling it on Windows so I will need to stop doing that when we roll past this patchset). It would be really nice to keep full Windows support in dump_syms (I'm not sure if crashpad re-implemented that or not?).

>

Excellent, that would do nicely. I would be satisfied with applying a patch to our breakpad tree independently if you decide not to.

Thanks again, and keep up the good work. :)

Patch-set: 4
Tag: mailMessageId=<CACK9vu57DtF1JknWkK+bDQ8HqnBnZH0jqCbzbxswRpsdUugf3Q@mail.gmail.com>