Update patch set 4
Patch Set 4:

> Thanks very much for the feedback. My takeaway is that you will continue to support Windows generally which is a relief. stackwalk isn't so much of an issue for me as we do that server-side (although we are currently compiling it on Windows so I will need to stop doing that when we roll past this patchset). It would be really nice to keep full Windows support in dump_syms (I'm not sure if crashpad re-implemented that or not?).

Only minidump_dump and minidump_stackwalk were converted to use getopt(). Those are both Breakpad processor tools. Traditionally, Windows support for the Breakpad processor library and tools has been hit-or-miss. They often work, but it’s not a top priority, and occasionally things (like the recent getopt() and basename()) slip in that prevent them from working on Windows. There’s nothing intentional about this, we’re just not testing or monitoring that configuration. I also don’t actually know whether this code built on Windows prior to these recent changes. Note that this lack of first-class Windows support only refers to the system that runs the processor library and tools. The processor is always fully able to interpret Windows crashes regardless of the platform that it runs on.

Distinctly, dump_syms is not a processor tool at all, it’s a build tool. Because we use the Microsoft DIA SDK to extract symbols from Windows PE/COFF images, this tool must run on Windows, since that’s the only place that DIA runs. As such, as long as we’re using the Breakpad processor library to deal with crashes, and we need to continue using the Breakpad dumped symbol format (since that’s what the processor library expects), we’ll need to ensure that dump_syms on Windows functions.

Crashpad has obsoleted the Windows (and macOS) Breakpad clients. We are working on a new processor system that will eventually replace Breakpad’s, but it’s not up and running yet, and so even though we’re using Crashpad to produce crash dumps on two platforms, it’s still providing these dumps to Breakpad-type collection servers where a Breakpad processor makes sense of them using symbol files in Breakpad’s dumped format. None of that’s changed yet, and we certainly wouldn’t do anything to interfere with dump_syms running on Windows as it’s an integral part of that pipeline.

> Excellent, that would do nicely. I would be satisfied with applying a patch to our breakpad tree independently if you decide not to.
> 
> Thanks again, and keep up the good work. :)

Patch-set: 4