Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> There are other places where we assume that string == std::string.
> 
> In fact, we have a header, common/using_std_string.h, that’s supposed to do the right thing. Is that not #included here?

I looked into this, and the code doesn't actually assume that string == std::string, but merely that there are implicit conversions between the two types.

The common/using_std_string.h header has no effect when both ::string and std::string are defined and HAS_GLOBAL_STRING is defined. All it does is copy std::string into the global namespace when HAS_GLOBAL_STRING is not defined.

The reason for the compilation error is that std::map<std::string, std::string> is a different type than std::map<::string, ::string> when both strings are defined, so an iterator of one cannot be assigned to a variable of the iterator type of the other.

The options I see to fix this are either this change, which seems to be consistent with the other STL type references in the file, or changing the type definitions in the header to use string.

Patch-set: 1
Reviewer: Mike Wittman <1002120@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>