Update patch set 1
Patch Set 1:

Furthermore, it’d be good to #include_next headers from the actual NDK if they’re present, and only fall back to our fill-ins if we don’t find NDK headers or know that they won’t be there or won’t be suitable.

You may be able to use __has_include_next or __ANDROID_API__ as building blocks for detecting what’s available or possible. In cases where a header is guaranteed to be present but may not have the content you’re looking for, the best option is normally to #include_next blindly and then test a macro that the genuine header #defines.

Patch-set: 1
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>