Update patch set 2
Patch Set 2:

here's the current gen+upload from a random bot:

https://uberchromegw.corp.google.com/i/chromeos/builders/link-release/builds/6678 -> DebugSymbols

> 21:43:48: INFO: generating breakpad symbols using /build/link/usr/lib/debug
> 22:04:26: INFO: Dumped /build/link/opt/google/chrome/chrome as chrome : 4BC245817721C8F6573CDF1D816ED7BE0
> 22:05:26: INFO: Announcing availability of debug tarball now.

Chrome by far took the longest to generate (~20 min).

> 22:24:52: INFO: RunCommand: /b/c/cbuild/repository/chromite/bin/upload_symbols --yes --dedupe --root /b/c/cbuild/repository/chroot --board link --official_build --failed-list /b/c/cbuild/repository/buildbot_archive/link-release/R69-10734.0.0/failed_upload_symbols.list in /b/c/cbuild/repository
> 22:34:18: INFO: RunCommand: /b/c/cbuild/repository/.cache/common/gsutil_4.30.tar.gz/gsutil/gsutil -o 'Boto:num_retries=10' -m cp -v -- /b/c/cbuild/repository/buildbot_archive/link-release/R69-10734.0.0/failed_upload_symbols.list gs://chromeos-image-archive/link-release/R69-10734.0.0/failed_upload_symbols.list

the upload took less time because of our use of swarming for deduping symbols across bots.  but if we had to upload all the files, it would indeed have taken much longer.

iirc, we used to run the gen+upload in parallel to amortize the costs of both, but now we just use swarming to dedupe after generation as that cuts down on load on the crash servers too.

Patch-set: 2
CC: Mike Frysinger <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>