fix(build): Remove libdis namespace to fix C++ module import error
With C++ modules enabled (`-fmodules`), including a system header like
<stdint.h> from within a namespace is not permitted. This was causing a
build failure in the `:stackwalk_common` target:

  fatal error: import of module '_Builtin_stdint' appears within
  namespace 'libdis'

The code in `disassembler_x86.h` was wrapping the `libdisasm`
third-party library, including its headers, inside a `libdis` namespace.

This change removes the `libdis` namespace wrapper, resolving the build
failure. This allows C++ modules to be re-enabled for the breakpad
target, contributing to the broader effort to use modules for system
headers.

Bug: 424364315
Change-Id: I9e5bfe29f63b6e1659c7912a15a3eaa376b865d2
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/6636461
Reviewed-by: Primiano Tucci <primiano@chromium.org>