Update dump_syms to use C++20
This change updates the dump_syms Visual Studio project to support
C++20 [1]. This is needed for an upcoming CL that adds a use of
std::optional (part of C++17).

Breakpad requires C++20 as of earlier this year [2], so this CL makes
dump_syms' VS project match.

This change also removes option `/Gm` (minimal rebuild), which is
deprecated and causes the following warning (which becomes an error
under C++20):
  "D9035: option 'Gm' has been deprecated and will be removed in a
  future release"

[1] In Visual Studio settings, this corresponds to changing
    "Properties > C/C++ > Language > C++ Language Standard"
    from "Default (ISO C++14 Standard)"
    to "ISO C++20 Standard (/std:c++20)"
[2] [6347230: increase C++ standard to 20](https://chromium-review.googlesource.com/c/breakpad/breakpad/+/6347230)

Change-Id: Ie19bc85d8148bd5004a60a9957be9d9cdb0b3ce6
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/7205053
Reviewed-by: Mike Frysinger <vapier@chromium.org>