Allow symupload to take .sym files as input.
This change enables symupload to directly process Breakpad symbol files (.sym). When a .sym file is provided, the module information (OS, CPU, debug ID, and filename) is extracted from the first line of the .sym file. The temporary file cleanup is also adjusted to only remove the generated .sym file when the input was not already a .sym file.

Bug: 459506875
Change-Id: I54c796733654aebf30528ac6eda6f0ca5b85378a
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/7231157
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Joshua Peraza <jperaza@chromium.org>